Updated to "Evasion Pruning Tweak". https://github.com/official-stockfish/Stockfish/commit/1d31065e1d1f04ecf47fc959523c560e8657fbfa

+++ src/movepick.h
@@ -35,7 +35,7 @@ INLINE void cms_update(CounterMoveStats cms, Piece pc, Square to, int v)
   int w = v >= 0 ? v : -v;
 
   cms[pc][to] -= cms[pc][to] * w / 936;
-  cms[pc][to] += ((int)v) * 32;
+  cms[pc][to] += v * 32;
 }
 
 INLINE void history_update(HistoryStats history, int c, Move m, int v)
@@ -44,7 +44,7 @@ INLINE void history_update(HistoryStats history, int c, Move m, int v)
 
   m &= 4095;
   history[c][m] -= history[c][m] * w / 324;
-  history[c][m] += ((int)v) * 32;
+  history[c][m] += v * 32;
 }
 
 INLINE int history_get(HistoryStats history, int c, Move m)
